Baseline EnDeLSTM
baseline model trained on one-hot encoded notes. next step is to stack lstms so velocity and delta time depend on the not being played. This is an ambitious effort in reality. still have lto figure how to train 3 connected models with different inputs and utputs with a common loss function and optimizer.